refactor: move createExpressApp to server file and simplify run file
